+++ desktop/libs/liboozie/src/liboozie/submission2.py
@@ -15,16 +15,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from builtins import object
-import errno
-import logging
 import os
 import sys
 import time
-
+import errno
+import logging
+from builtins import object
 from string import Template
 
 from django.utils.functional import wraps
+from django.utils.translation import gettext as _
 
 from beeswax.hive_site import get_hive_site_content
 from desktop.lib.exceptions_renderable import PopupException
@@ -32,23 +32,16 @@ from desktop.lib.i18n import smart_str
 from desktop.lib.parameterization import find_variables
 from desktop.lib.paths import get_desktop_root
 from desktop.models import Document2
-from indexer.conf import CONFIG_JDBC_LIBS_PATH
-from metadata.conf import ALTUS
-from oozie.utils import convert_to_server_timezone
-
 from hadoop import cluster
 from hadoop.fs.hadoopfs import Hdfs
-
+from indexer.conf import CONFIG_JDBC_LIBS_PATH
 from liboozie.conf import REMOTE_DEPLOYMENT_DIR, USE_LIBPATH_FOR_JARS
 from liboozie.credentials import Credentials
 from liboozie.oozie_api import get_oozie
+from metadata.conf import ALTUS
+from oozie.utils import convert_to_server_timezone
 
-if sys.version_info[0] > 2:
-  from django.utils.translation import gettext as _
-else:
-  from django.utils.translation import ugettext as _
-
-LOG = logging.getLogger(__name__)
+LOG = logging.getLogger()
 
 
 def submit_dryrun(run_func):
@@ -96,7 +89,7 @@ class Submission(object):
       local_tz = self.job.data.get('properties')['timezone']
 
     # Modify start_date & end_date only when it's a coordinator
-    from oozie.models2 import Coordinator, Bundle
+    from oozie.models2 import Bundle, Coordinator
     if type(self.job) is Coordinator:
       if 'start_date' in self.properties:
         properties['start_date'] = convert_to_server_timezone(self.properties['start_date'], local_tz)
@@ -292,8 +285,8 @@ python altus.py
               self.fs.do_as_user(self.user, self.fs.copyFromLocal, os.path.join(source_path, name), destination_path)
 
         elif action.data['type'] == 'impala' or action.data['type'] == 'impala-document':
-          from oozie.models2 import _get_impala_url
           from impala.impala_flags import get_ssl_server_certificate
+          from oozie.models2 import _get_impala_url
 
           if action.data['type'] == 'impala-document':
             from notebook.models import Notebook
@@ -408,10 +401,12 @@ STORED AS TEXTFILE %s""" % (self.properties.get('send_result_path'),
     """From XML and job.properties HDFS files"""
     deployment_dir = os.path.dirname(application_path)
     xml = self.fs.do_as_user(self.user, self.fs.read, application_path, 0, 1 * 1024 ** 2)
+    xml = xml.decode('utf-8') if hasattr(xml, 'decode') else xml
 
     properties_file = deployment_dir + '/job.properties'
     if self.fs.do_as_user(self.user, self.fs.exists, properties_file):
       properties = self.fs.do_as_user(self.user, self.fs.read, properties_file, 0, 1 * 1024 ** 2)
+      properties = properties.decode('utf-8') if hasattr(properties, 'decode') else properties
     else:
       properties = None
 
@@ -555,7 +550,10 @@ STORED AS TEXTFILE %s""" % (self.properties.get('send_result_path'),
         jar_path = node.data['properties'].get('jar_path')
         if jar_path:
           if not jar_path.startswith('/'):  # If workspace relative path
-            jar_path = self.fs.join(self.job.deployment_dir, jar_path)
+            if jar_path.startswith('s3a://') or jar_path.startswith('abfs://'):
+              jar_path = jar_path
+            else:
+              jar_path = self.fs.join(self.job.deployment_dir, jar_path)
           if not jar_path.startswith(lib_path):  # If not already in lib
             files.append(jar_path)
 
@@ -572,12 +570,15 @@ STORED AS TEXTFILE %s""" % (self.properties.get('send_result_path'),
       if files:
         for jar_file in files:
           LOG.debug("Updating %s" % jar_file)
-          jar_lib_path = self.fs.join(lib_path, self.fs.basename(jar_file))
+          if jar_file.startswith('s3a://') or jar_file.startswith('abfs://'):
+            jar_lib_path = jar_file
+          else:
+            jar_lib_path = self.fs.join(lib_path, self.fs.basename(jar_file))
           # Refresh if needed
           if self.fs.exists(jar_lib_path) and self.fs.exists(jar_file):
             stat_src = self.fs.stats(jar_file)
             stat_dest = self.fs.stats(jar_lib_path)
-            if stat_src.fileId != stat_dest.fileId:
+            if hasattr(stat_src, 'fileId') and hasattr(stat_dest, 'fileId') and stat_src.fileId != stat_dest.fileId:
               self.fs.remove(jar_lib_path, skip_trash=True)
           self.fs.copyfile(jar_file, jar_lib_path)
 
@@ -608,10 +609,16 @@ STORED AS TEXTFILE %s""" % (self.properties.get('send_result_path'),
 
   def _create_file(self, deployment_dir, file_name, data, do_as=False):
     file_path = self.fs.join(deployment_dir, file_name)
+
+    # In Py3 because of i18n, the xml data is not properly utf-8 encoded for some languages.
+    # This can later throw UnicodeEncodeError exception for request body in HDFS or other FS API calls. To tackle this,
+    # We are converting the data into bytes by utf-8 encoding instead of str type.
+    data = smart_str(data).encode('utf-8')
+
     if do_as:
-      self.fs.do_as_user(self.user, self.fs.create, file_path, overwrite=True, permission=0o644, data=smart_str(data))
+      self.fs.do_as_user(self.user, self.fs.create, file_path, overwrite=True, permission=0o644, data=data)
     else:
-      self.fs.create(file_path, overwrite=True, permission=0o644, data=smart_str(data))
+      self.fs.create(file_path, overwrite=True, permission=0o644, data=data)
     LOG.debug("Created/Updated %s" % (file_path,))
 
   def _generate_altus_action_script(self, service, command, arguments, auth_key_id, auth_key_secret):
@@ -665,7 +672,7 @@ print _exec('%(service)s', '%(command)s', %(args)s)
     else:
       hostname = ALTUS.HOSTNAME.get()
 
-    if type(cluster) == dict:
+    if type(cluster) is dict:
       command = 'createAWSCluster'
       arguments = cluster
     else:
